backend/foodgram_backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.16.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.8.1
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
